Preparing custom SQL queries
This may not have been strictly necessary, since it's reasonable to assume that $wpdb->posts, $event_ids, etc are safe, but it's better to err on the side of caution.

It also helps to avoid human-error in the future, if an inexperienced programmer were to modify the statement. Because it's now already setup to prepare() data, it's much more likely that they would prepare() their new variable, whereas it's likely they would otherwise fail to realize that the statement now is unsafe and needs to be prepare()'d.