No longer passing two SQL statements through prepare()
They were added in a previous commit even though the data was safe, as a best-practice to future proof the statement. The event_ids variable can't be treated as a string, though, so it needs to be included inline. Calling prepare() with no parameters triggers a doing_it_wrong() notice, so it's ok to leave it unprepared.