Removing nonce from log search form
Having the nonce made it impossible to deep-link to the log entry in e-mail notifications, since the nonce would be generated by the user whose action triggered the notifcation, but the link that included the nonce was intended to be viewed by admins. Since the nonce wasn't assigned to the admins, it would fail verification.

Having nonces on a search form is probably overkill anyway, since it's only returning data, rather than making changes to the system state.