Prevent revenue report cron job from running more often than intended
Because the plugin is network activated, the cron job is scheduled for each site on the network. Without any logic to prevent it, this results in the job being fired once for every site at each interval, rather than just once per interval for the entire network.

This patch ensures the job will only run once on the network at each interval by checking how long it's been since the previous run, and returning early if it's not time for the next run yet.